requirements.txt
boto3>=1.34.0
requests>=2.31.0
beautifulsoup4>=4.12.0
flask>=3.0.0
python-dotenv>=1.0.0
